Separate header files from Sources folder
